requirements.txt
gradio
langchain
langchain-openai
python-dotenv
tenacity
pydantic
